Fix spotless cache miss
Spotless custom steps are never up to date by default.
This is achieved using a randomly generated number attached to the custom step's state, which causes the custom step to never be up-to-date between runs.
Calling bumpThisNumberIfACustomStepChanges(int) pins this randomly generated number to the provided int, but it must be updated if a custom step changes, otherwise incorrect cache hits are possible.